requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0